Using dbs to replace db dialog in trust.
TODO: Interpreter is not good for choose which round.
I should write interpreted code into string and save'
to a table with round number.  When need, take the
string and run as a program.

The launch is not as fast as possible.  I should use
the manualTouch to do the launch.